src/serving/requirements.txt
google-cloud-storage==2.7.0
flask>=2.2
joblib==1.2.0
scikit-learn>=1.0
numpy>=1.21
pandas>=1.0
